requirements.txt
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
cssselect2==0.7.0
distlib==0.3.6
Django==4.2.3
django-admin-persian-fonts==0.2
django-bootstrap-v5==1.0.11
django-colorfield==0.9.0
django-cors-headers==3.4.0
django-debug-toolbar==2.2
django-extensions==3.0.3
django-filter==23.2
django-jalali==6.0.1
django-jalali-date==1.0.2
django-redis==5.2.0
django-weasyprint==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
filelock==3.8.0
fonttools==4.39.4
gunicorn==20.0.4
html5lib==1.1
idna==2.10
inflection==0.5.1
isodate==0.6.1
jdatetime==4.1.1
lxml==4.9.2
Markdown==3.2.2
openpyxl==3.1.2
packaging==23.2
persiantools==3.0.1
Pillow==9.5.0
platformdirs==2.5.2
psycopg2==2.9.5
pycparser==2.21
pydot==1.4.1
pydyf==0.6.0
PyJWT==2.7.0
pyparsing==2.4.7
pyphen==0.14.0
python-memcached==1.59
pytz==2020.1
PyYAML==6.0.1
redis==4.5.5
requests==2.23.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rest-condition==1.0.3
six==1.15.0
soupsieve==2.4.1
sqlparse==0.3.1
tinycss2==1.2.1
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.25.11
validate-email==1.3
virtualenv==20.16.6
weasyprint==59.0
webencodings==0.5.1
xlwt==1.3.0
zeep==4.2.1
zopfli==0.2.2
